extracted methods
