+++ src/main/java/lt/liutikas/reddit/controller/delegate/DefaultNewsApiDelegate.java
@@ -2,7 +2,6 @@ package lt.liutikas.reddit.controller.delegate;
 
 import lt.liutikas.reddit.api.model.GetNewsRequest;
 import lt.liutikas.reddit.api.model.NewsPage;
-import lt.liutikas.reddit.api.model.SaveNewsRequest;
 import lt.liutikas.reddit.assembler.NewsAssembler;
 import lt.liutikas.reddit.model.core.Channel;
 import lt.liutikas.reddit.openapi.api.NewsApiDelegate;
@@ -13,8 +12,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -40,16 +37,8 @@ public class DefaultNewsApiDelegate implements NewsApiDelegate {
 
     @Override
     public ResponseEntity<News> createNews(CreateNewsRequest createNewsRequest) {
-        SaveNewsRequest request = new SaveNewsRequest();
-        request.setSubChannel(createNewsRequest.getSubChannel());
-        request.setTitle(createNewsRequest.getTitle());
-        request.setChannel(Channel.valueOf(createNewsRequest.getChannel().getValue()));
-        request.setUrl(parseUrl(createNewsRequest));
-
-        lt.liutikas.reddit.model.core.News news = newsService.saveNews(request);
-        News openApiNews = newsAssembler.assembleNews(news);
-
-        return new ResponseEntity<>(openApiNews, HttpStatus.CREATED);
+        News news = newsService.createNews(createNewsRequest);
+        return new ResponseEntity<>(news, HttpStatus.CREATED);
     }
 
     @Override
@@ -75,13 +64,6 @@ public class DefaultNewsApiDelegate implements NewsApiDelegate {
         return new ResponseEntity<>(openApiNewsPage, HttpStatus.OK);
     }
 
-    private URL parseUrl(CreateNewsRequest createNewsRequest) {
-        try {
-            return new URL(createNewsRequest.getUrl());
-        } catch (MalformedURLException e) {
-            throw new IllegalArgumentException(String.format("\"%s\" is not a valid url", createNewsRequest.getUrl()), e);
-        }
-    }
 
     private lt.liutikas.reddit.openapi.model.NewsPage getOpenApiNewsPage(NewsPage newsPage) {
         lt.liutikas.reddit.openapi.model.NewsPage openApiNewsPage = new lt.liutikas.reddit.openapi.model.NewsPage();
